data/requirements/requirements_dev.txt
-r requirements.txt
requests==2.28.2
python-dotenv==1.0.0